% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddna.R
\name{longest_common_subsequence}
\alias{longest_common_subsequence}
\title{Title}
\usage{
longest_common_subsequence(df)
}
\arguments{
\item{array}{of dna sequence}
}
\value{
print
}
\description{
Title
}
\examples{
X <- c('AAAABBAABABBB', 'AAABAABABBB', 'BAAABAABABBA', 'ABAABBBAB')
longest_common_subsequence(array)
}
