% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddna.R
\name{sequence_tweets_from_file}
\alias{sequence_tweets_from_file}
\title{Title}
\usage{
sequence_tweets_from_file(input_file, alphabet = "b3_type")
}
\arguments{
\item{alphabet}{}
}
\value{
stt
}
\description{
Title
}
\examples{

}
